fix(py): race condition on rule propagation
Policy enforcement happens on top of policy rules, which are records resulting from decomposing a policy and propagating fragments through the DDB stream and towards an index. Stream propagation and indexing are both asynchronous operations, and as a result, policy fragments that are not absolute could be available to support enforcement, possibly resulting in security flaws. That would happen, for example, when a policy containing an allowance for a certain action also contains a conditional denial for that same action, and then, because of eventual consistency streaming, only the allowance rule ends up available in the enforcement index for a period of time. This commit fixes that problem by augmenting the propagated fragments with metadata that allows enforcement code to verify whether rule set consistency has been achieved in the index.